backend_v2.0/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
aiomysql
redis
pydantic
